Type fix

+++ PrettyNumberCounter.h
@@ -20,7 +20,7 @@ private:
 		}
 		for (size_t step : std::views::iota(1u, region_size)) {
 			for (size_t sum : std::views::iota(0u, (num_system_base - 1) * (step + 1) + 1)) {
-				size_t subsum = 0;
+				uint32_t subsum = 0;
 				for (size_t digit : std::views::iota(0u, std::min(sum + 1, num_system_base))) {
 					subsum += sums[max_sum * (step - 1) + sum - digit];
 				}
@@ -33,7 +33,7 @@ private:
 	{
 		uint64_t result = 0;
 		for (size_t sum : std::views::iota(0u, max_sum)) {
-			result += sums[max_sum * (region_size - 1) + sum] * sums[max_sum * (region_size - 1) + sum];
+			result += static_cast<uint64_t>(sums[max_sum * (region_size - 1) + sum]) * sums[max_sum * (region_size - 1) + sum];
 		}
 		for (size_t unused_digit : std::views::iota(0u, unused_digits)) {
 			result *= num_system_base;
